requirements.txt
crewai~=0.126.0
langchain-google-genai~=2.1.5
PyYAML~=6.0.2
requests~=2.32.4
beautifulsoup4~=4.13.4
python-dotenv~=1.1.0